Refactor medication and inventory data types for improved precision
- Changed `doseCountPerIntake`, `inventoryCount`, and `inventoryQuantity` fields in the `Medication` model to `Float` for better handling of decimal values.
- Updated `quantityTaken` in `PrnDoseRecord` and `delta` in `MedicationInventoryAdjustment` to `Float` to accommodate fractional quantities.
- Adjusted related validation logic in `inventory.ts` and `prnDoseRecord.ts` to accept `Float` values instead of integers.
- Modified various DTOs and view models to reflect these changes, ensuring consistent data handling across the application.
- Enhanced UI components to display decimal values correctly, improving user experience in medication and inventory management.

+++ ios/MedicationApp/Features/Caregiver/CaregiverEventSubscriber.swift
@@ -16,7 +16,7 @@ struct InventoryAlertEvent: Equatable, Decodable {
     let patientId: String
     let medicationId: String
     let type: String
-    let remaining: Int
+    let remaining: Double
     let threshold: Int
     let patientDisplayName: String?
     let medicationName: String?
